+++ zenith-float-num/tests/mpfr/mod.rs
@@ -1,5 +1,7 @@
 mod common;
+pub mod compare_complex_test;
 pub mod compare_const_test;
 pub mod compare_ops_test;
+pub mod compare_special_fn_test;
 pub mod compare_special_test;
 pub mod fuzz_round_modes;